Refatora scraping de BDR em parsers e coletor dedicados
